server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
motor==3.5.0
pymongo==4.8.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
pydantic==2.9.0
pydantic-settings==2.5.0
python-dotenv==1.0.1
redis[hiredis]==5.1.0
slowapi==0.1.9
httpx==0.27.0
